Update README.md
Today I'm having trouble pushing my commits to our GitHub repositories the usual way (via SourceTree) and the not-so-usual way (via git commands on unix). Fortunately, there's a way for me to update this README file directly here on GitHub. There's a lot of stuff to learn and do everyday. Thank you. Peace.